+++ weblate/addons/autotranslate.py
@@ -80,8 +80,12 @@ class AutoTranslateAddon(BaseAddon):
                 source_component_id=conf["component"],
             )
 
-    def daily(self, component: Component, activity_log_id: int | None = None) -> None:
-        # Translate every component less frequenctly to reduce load.
+    def daily_component(
+        self,
+        component: Component,
+        activity_log_id: int | None = None,
+    ) -> None:
+        # Translate every component less frequently to reduce load.
         # The translation is anyway triggered on update, so it should
         # not matter that much that we run this less often.
         if settings.BACKGROUND_TASKS == "never":